feat(crons): Ensure that org level monitor endpoints will continue working if there are multiple monitors with the same slug
Since we're making the constraint for monitor slug on (project, slug) we might end up with multiple potential monitors in the org level endpoints. This puts in a temporary fix so that we just return the oldest monitor.

I can't write a test for this yet since we still have the old constraint in place.